Module Apronext__.Domain

module type ADomain = sig ... end

Abstract domain signature used by the following functor

module Make : functor (D : ADomain) -> sig ... end

functor that allows hiding the use of the manager, It also adds few utilities to the Abstract1 module