Module Apronext.Scalarext

include Apron.Scalar

type t = Apron.Scalar.t =

| Float of float

| Mpqf of Mpqf.t

| Mpfrf of Mpfrf.t

val of_mpq : Mpq.t -> t

val of_mpqf : Mpqf.t -> t

val of_int : int -> t

val of_frac : int -> int -> t

val of_mpfr : Mpfr.t -> t

val of_mpfrf : Mpfrf.t -> t

val of_float : float -> t

val of_infty : int -> t

val is_infty : t -> int

val sgn : t -> int

val cmp : t -> t -> int

val cmp_int : t -> int -> int

val equal : t -> t -> bool

val equal_int : t -> int -> bool

val neg : t -> t

val to_string : t -> string

val print : Stdlib.Format.formatter -> t -> unit

val to_mpqf : t -> Mpqf.t

val to_float : t -> float

val add : t -> t -> t

scalar addition. result is automatically lifted to mpqf to avoid loss of precision

val sub : t -> t -> t

scalar substraction. result is automatically lifted to mpqf to avoid loss of precision

val mul : t -> t -> t

scalar multiplication. result is automatically lifted to mpqf to avoid loss of precision

val div : t -> t -> t

scalar division. result is automatically lifted to mpqf to avoid loss of precision